+++ src/org/authInfo.ts
@@ -372,14 +372,16 @@ export class AuthInfo extends AsyncOptionalCreatable<AuthInfo.Options> {
    * ```
    * @param sfdxAuthUrl
    */
-  public static parseSfdxAuthUrl(sfdxAuthUrl: string) {
+  public static parseSfdxAuthUrl(
+    sfdxAuthUrl: string
+  ): Pick<AuthFields, 'clientId' | 'clientSecret' | 'refreshToken' | 'loginUrl'> {
     const match = sfdxAuthUrl.match(
       /^force:\/\/([a-zA-Z0-9._-]+):([a-zA-Z0-9._-]*):([a-zA-Z0-9._-]+={0,2})@([a-zA-Z0-9._-]+)/
     );
 
     if (!match) {
       throw new SfdxError(
-        'Invalid sfdx auth url. Must be in the format `force://<clientId>:<clientSecret>:<refreshToken>@<loginUrl>`. The instanceUrl must not have the protocol set.',
+        'Invalid SFDX auth URL. Must be in the format "force://<clientId>:<clientSecret>:<refreshToken>@<instanceUrl>".  Note that the SFDX auth URL uses the "force" protocol, and not "http" or "https".  Also note that the "instanceUrl" inside the SFDX auth URL doesn\'t include the protocol ("https://").',
         'INVALID_SFDX_AUTH_URL'
       );
     }
@@ -1005,5 +1007,7 @@ export namespace AuthInfo {
      * creation.
      */
     parentUsername?: string;
+
+    isDevHub?: boolean;
   }
 }